Verif algo ok

+++ srcs/ft_event.c
@@ -16,17 +16,17 @@ int		ft_event(int keycode, t_scene *scn)
 
 int		ft_event_mouse(int button, int x, int y, t_scene *scn)
 {
-	double 	h;
+	float 	h;
 
 	h = 1.1;
 	if (button == 5)
 	{
 		printf("before %.20f\n", scn->f->x1);
 		scn->f->zoom *= h;
- 		scn->f->x1 += (double)x / scn->f->zoom / 10;
- 		scn->f->x2 += (double)x / scn->f->zoom / 10;
- 		scn->f->y1 += (double)y / scn->f->zoom / 10;
- 		scn->f->y2 += (double)y / scn->f->zoom / 10;
+ 		scn->f->x1 += (float)x / scn->f->zoom / 10;
+ 		scn->f->x2 += (float)x / scn->f->zoom / 10;
+ 		scn->f->y1 += (float)y / scn->f->zoom / 10;
+ 		scn->f->y2 += (float)y / scn->f->zoom / 10;
 		scn->f->ite_max += 1;
 		printf("after %.20f %d %f\n", scn->f->x1, scn->f->ite_max, scn->f->zoom);
 	}
@@ -34,10 +34,10 @@ int		ft_event_mouse(int button, int x, int y, t_scene *scn)
 	{
 		printf("- before %.20f\n", scn->f->x1);
 		scn->f->zoom /= h;
- 		scn->f->x1 -= (double)x / scn->f->zoom / 10;
- 		scn->f->x2 -= (double)x / scn->f->zoom / 10;
- 		scn->f->y1 -= (double)y / scn->f->zoom / 10;
- 		scn->f->y2 -= (double)y / scn->f->zoom / 10;
+ 		scn->f->x1 -= (float)x / scn->f->zoom / 10;
+ 		scn->f->x2 -= (float)x / scn->f->zoom / 10;
+ 		scn->f->y1 -= (float)y / scn->f->zoom / 10;
+ 		scn->f->y2 -= (float)y / scn->f->zoom / 10;
 		scn->f->ite_max -= 1;
 		printf("- after %.20f\n", scn->f->x1);
 	}
